+++ packages/braintree-utils/src/create-braintree-sdk.spec.ts
@@ -1,8 +1,15 @@
+import { PaymentIntegrationService } from '@bigcommerce/checkout-sdk/payment-integration-api';
+import { PaymentIntegrationServiceMock } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
+
 import BraintreeSdk from './braintree-sdk';
 import createBraintreeSdk from './create-braintree-sdk';
 
 describe('createBraintreeSdk', () => {
+    let paymentIntegrationService: PaymentIntegrationService;
+
     it('instantiates braintree sdk', () => {
-        expect(createBraintreeSdk()).toBeInstanceOf(BraintreeSdk);
+        paymentIntegrationService = new PaymentIntegrationServiceMock();
+
+        expect(createBraintreeSdk(paymentIntegrationService)).toBeInstanceOf(BraintreeSdk);
     });
 });