+++ packages/paypal-commerce-utils/src/create-paypal-commerce-fastlane-utils.ts
@@ -1,7 +1,5 @@
-import { BrowserStorage } from '@bigcommerce/checkout-sdk/storage';
-
 import PayPalCommerceFastlaneUtils from './paypal-commerce-fastlane-utils';
 
 export default function createPayPalCommerceFastlaneUtils(): PayPalCommerceFastlaneUtils {
-    return new PayPalCommerceFastlaneUtils(new BrowserStorage('paypalFastlane'));
+    return new PayPalCommerceFastlaneUtils();
 }