+++ packages/paypal-commerce-utils/src/paypal-commerce-types.ts
@@ -452,7 +452,7 @@ export interface PayPalFastlaneAuthenticationResult {
 export enum PayPalFastlaneAuthenticationState {
     SUCCEEDED = 'succeeded',
     FAILED = 'failed',
-    CANCELED = 'cancelled',
+    CANCELED = 'canceled',
     UNRECOGNIZED = 'unrecognized',
 }