+++ packages/stripe-integration/src/stripe-ocs/stripe-ocs-payment-strategy.ts
@@ -29,7 +29,9 @@ import {
     StripeError,
     StripeEventType,
     StripeInitializationData,
+    StripeInstrumentSetupFutureUsage,
     StripeIntegrationService,
+    StripePIPaymentMethodOptions,
     StripeResult,
     StripeScriptLoader,
     StripeStringConstants,
@@ -121,7 +123,10 @@ export default class StripeOCSPaymentStrategy implements PaymentStrategy {
     }
 
     deinitialize(): Promise<void> {
-        this.stripeElements?.getElement(StripeElementType.PAYMENT)?.unmount();
+        const paymentElement = this.stripeElements?.getElement(StripeElementType.PAYMENT);
+
+        paymentElement?.unmount();
+        paymentElement?.destroy();
         this.stripeIntegrationService.deinitialize();
         this.stripeElements = undefined;
         this.stripeClient = undefined;
@@ -151,7 +156,7 @@ export default class StripeOCSPaymentStrategy implements PaymentStrategy {
         }
 
         const { clientToken, initializationData } = paymentMethod;
-        const { shopperLanguage, customerSessionToken } = initializationData;
+        const { shopperLanguage, customerSessionToken, enableLink } = initializationData;
 
         if (!clientToken) {
             throw new MissingDataError(MissingDataErrorType.MissingPaymentMethod);
@@ -178,7 +183,8 @@ export default class StripeOCSPaymentStrategy implements PaymentStrategy {
         });
 
         const { getBillingAddress, getShippingAddress } = this.paymentIntegrationService.getState();
-        const { postalCode } = getShippingAddress() || getBillingAddress() || {};
+        const billingAddress = getBillingAddress();
+        const { postalCode } = getShippingAddress() || billingAddress || {};
 
         const stripeElement: StripeElement =
             this.stripeElements.getElement(StripeElementType.PAYMENT) ||
@@ -198,10 +204,16 @@ export default class StripeOCSPaymentStrategy implements PaymentStrategy {
                 wallets: {
                     applePay: StripeStringConstants.NEVER,
                     googlePay: StripeStringConstants.NEVER,
+                    link: enableLink ? StripeStringConstants.AUTO : StripeStringConstants.NEVER,
                 },
                 layout,
                 savePaymentMethod: {
                     maxVisiblePaymentMethods: 20,
+                },
+                defaultValues: {
+                    billingDetails: {
+                        email: billingAddress?.email || '',
+                    },
                 },
             });
 
@@ -234,13 +246,18 @@ export default class StripeOCSPaymentStrategy implements PaymentStrategy {
         stripeElement?.collapse();
     }
 
-    private _getPaymentPayload(methodId: string, token: string): Payment {
+    private _getPaymentPayload(
+        methodId: string,
+        token: string,
+        shouldSaveInstrument = false,
+    ): Payment {
         const cartId = this.paymentIntegrationService.getState().getCart()?.id || '';
         const formattedPayload = {
             cart_id: cartId,
             credit_card_token: { token },
             confirm: false,
             payment_method_id: this.selectedMethodId,
+            vault_payment_instrument: shouldSaveInstrument,
         };
 
         return {
@@ -273,8 +290,15 @@ export default class StripeOCSPaymentStrategy implements PaymentStrategy {
             methodId,
             additionalActionData,
         );
+        const { id: paymentIntentId, payment_method_options: paymentMethodOptions } =
+            paymentIntent || {};
 
-        const paymentPayload = this._getPaymentPayload(methodId, paymentIntent?.id || token);
+        const shouldSaveCard = this._shouldSaveCard(paymentMethodOptions);
+        const paymentPayload = this._getPaymentPayload(
+            methodId,
+            paymentIntentId || token,
+            shouldSaveCard,
+        );
 
         try {
             return await this.paymentIntegrationService.submitPayment(paymentPayload);
@@ -328,5 +352,14 @@ export default class StripeOCSPaymentStrategy implements PaymentStrategy {
 
         this.selectedMethodId = event.value.type;
         paymentMethodSelect?.(`${gatewayId}-${methodId}`);
+    }
+
+    private _shouldSaveCard(paymentMethodOptions?: StripePIPaymentMethodOptions) {
+        const futureUsage = paymentMethodOptions?.card?.setup_future_usage;
+
+        return (
+            futureUsage === StripeInstrumentSetupFutureUsage.ON_SESSION ||
+            futureUsage === StripeInstrumentSetupFutureUsage.OFF_SESSION
+        );
     }
 }